refactor: consolidate verification handlers under /v1/channel-verification-sessions
