+++ assistant/src/runtime/routes/contact-routes.ts
@@ -7,12 +7,8 @@
  * DELETE /v1/contacts/:id          — delete a contact
  * POST   /v1/contacts/merge        — merge two contacts
  * PATCH  /v1/contact-channels/:contactChannelId — update a contact channel's status/policy
- * POST   /v1/contact-channels/:contactChannelId/verify — initiate trusted contact verification
  */
 
-import { createHash, randomBytes } from "node:crypto";
-
-import type { ChannelId } from "../../channels/types.js";
 import { resolveGuardianName } from "../../config/user-reference.js";
 import {
   deleteContact,
@@ -34,26 +30,8 @@ import type {
   ContactRole,
   ContactType,
 } from "../../contacts/types.js";
-import { getCredentialMetadata } from "../../tools/credentials/metadata-store.js";
-import {
-  countRecentSendsToDestination,
-  createOutboundSession,
-  updateSessionDelivery,
-} from "../channel-verification-service.js";
 import { httpError } from "../http-errors.js";
 import type { RouteDefinition } from "../http-router.js";
-import {
-  deliverVerificationSlack,
-  deliverVerificationTelegram,
-  DESTINATION_RATE_WINDOW_MS,
-  MAX_SENDS_PER_DESTINATION_WINDOW,
-  normalizeTelegramDestination,
-} from "../verification-outbound-actions.js";
-import {
-  composeVerificationSlack,
-  composeVerificationTelegram,
-  GUARDIAN_VERIFY_TEMPLATE_KEYS,
-} from "../verification-templates.js";
 
 function withGuardianNameOverride<
   T extends { role: string; displayName: string },
@@ -431,247 +409,6 @@ export async function handleUpdateContactChannel(
 }
 
 // ---------------------------------------------------------------------------
-// Channel verification
-// ---------------------------------------------------------------------------
-
-/** Session TTL in seconds (matches challenge TTL of 10 minutes). */
-const SESSION_TTL_SECONDS = 600;
-
-/**
- * Map a contact channel type to the verification ChannelId used by the
- * guardian service. Returns null for unsupported channel types.
- */
-function toVerificationChannel(channelType: string): ChannelId | null {
-  switch (channelType) {
-    case "phone":
-      return "voice";
-    case "telegram":
-      return "telegram";
-    case "slack":
-      return "slack";
-    default:
-      return null;
-  }
-}
-
-/**
- * Get the Telegram bot username from credential metadata.
- * Falls back to process.env.TELEGRAM_BOT_USERNAME.
- */
-function getTelegramBotUsername(): string | undefined {
-  const meta = getCredentialMetadata("telegram", "bot_token");
-  if (
-    meta?.accountInfo &&
-    typeof meta.accountInfo === "string" &&
-    meta.accountInfo.trim().length > 0
-  ) {
-    return meta.accountInfo.trim();
-  }
-  return process.env.TELEGRAM_BOT_USERNAME || undefined;
-}
-
-/**
- * POST /v1/contact-channels/:contactChannelId/verify
- *
- * Initiate trusted contact verification for a specific channel. Sends a
- * verification code via SMS, Telegram, Slack, or voice and returns session
- * info so the client can track the verification flow.
- */
-export async function handleVerifyContactChannel(
-  contactChannelId: string,
-  assistantId: string,
-): Promise<Response> {
-  const channel = getChannelById(contactChannelId);
-  if (!channel) {
-    return httpError(
-      "NOT_FOUND",
-      `Channel "${contactChannelId}" not found`,
-      404,
-    );
-  }
-
-  const contact = getContact(channel.contactId);
-  if (!contact) {
-    return httpError(
-      "NOT_FOUND",
-      `Contact "${channel.contactId}" not found`,
-      404,
-    );
-  }
-
-  // Already verified — no need to re-verify
-  if (channel.status === "active" && channel.verifiedAt != null) {
-    return httpError("CONFLICT", "Channel is already verified", 409);
-  }
-
-  const verificationChannel = toVerificationChannel(channel.type);
-  if (!verificationChannel) {
-    return httpError(
-      "BAD_REQUEST",
-      `Verification is not supported for channel type "${channel.type}"`,
-      400,
-    );
-  }
-
-  const destination = channel.address;
-  if (!destination) {
-    return httpError(
-      "BAD_REQUEST",
-      "Channel has no address to send verification to",
-      400,
-    );
-  }
-
-  // Normalize Telegram destinations so rate-limit lookups are consistent with
-  // guardian-outbound-actions (strips leading '@', lowercases handles).
-  const effectiveDestination =
-    verificationChannel === "telegram"
-      ? normalizeTelegramDestination(destination)
-      : destination;
-
-  // Rate limit check
-  const recentSendCount = countRecentSendsToDestination(
-    verificationChannel,
-    effectiveDestination,
-    DESTINATION_RATE_WINDOW_MS,
-  );
-  if (recentSendCount >= MAX_SENDS_PER_DESTINATION_WINDOW) {
-    return httpError(
-      "RATE_LIMITED",
-      "Too many verification attempts to this destination. Please try again later.",
-      429,
-    );
-  }
-
-  // --- Telegram verification ---
-  if (verificationChannel === "telegram") {
-    // Telegram with known chat ID: identity is already bound
-    if (channel.externalChatId) {
-      const sessionResult = createOutboundSession({
-        channel: verificationChannel,
-        expectedChatId: channel.externalChatId,
-        expectedExternalUserId: channel.externalUserId ?? undefined,
-        identityBindingStatus: "bound",
-        destinationAddress: effectiveDestination,
-        verificationPurpose: "trusted_contact",
-      });
-
-      const telegramBody = composeVerificationTelegram(
-        GUARDIAN_VERIFY_TEMPLATE_KEYS.TELEGRAM_CHALLENGE_REQUEST,
-        {
-          code: sessionResult.secret,
-          expiresInMinutes: Math.floor(SESSION_TTL_SECONDS / 60),
-        },
-      );
-
-      const now = Date.now();
-      const sendCount = 1;
-      updateSessionDelivery(sessionResult.sessionId, now, sendCount, null);
-      deliverVerificationTelegram(
-        channel.externalChatId,
-        telegramBody,
-        assistantId,
-      );
-
-      return Response.json({
-        ok: true,
-        verificationSessionId: sessionResult.sessionId,
-        expiresAt: sessionResult.expiresAt,
-        sendCount,
-      });
-    }
-
-    // Telegram handle only (no chat ID): bootstrap flow
-    const { ensureTelegramBotUsernameResolved } =
-      await import("../channel-invite-transports/telegram.js");
-    await ensureTelegramBotUsernameResolved();
-    const botUsername = getTelegramBotUsername();
-    if (!botUsername) {
-      return httpError(
-        "BAD_REQUEST",
-        "Telegram bot username is not configured. Set up the Telegram integration first.",
-        400,
-      );
-    }
-
-    const bootstrapToken = randomBytes(16).toString("hex");
-    const bootstrapTokenHash = createHash("sha256")
-      .update(bootstrapToken)
-      .digest("hex");
-
-    const sessionResult = createOutboundSession({
-      channel: verificationChannel,
-      identityBindingStatus: "pending_bootstrap",
-      destinationAddress: effectiveDestination,
-      bootstrapTokenHash,
-      verificationPurpose: "trusted_contact",
-    });
-
-    const telegramBootstrapUrl = `https://t.me/${botUsername}?start=gv_${bootstrapToken}`;
-
-    return Response.json({
-      ok: true,
-      verificationSessionId: sessionResult.sessionId,
-      expiresAt: sessionResult.expiresAt,
-      sendCount: 0,
-      telegramBootstrapUrl,
-    });
-  }
-
-  // --- Slack verification ---
-  if (verificationChannel === "slack") {
-    const slackUserId = channel.externalUserId ?? destination;
-
-    // Only claim identity is bound when we have at least one platform identifier
-    const hasIdentityBinding = Boolean(
-      channel.externalUserId || channel.externalChatId,
-    );
-    if (!hasIdentityBinding) {
-      return httpError(
-        "BAD_REQUEST",
-        "Slack verification requires an externalUserId or externalChatId for identity binding",
-        400,
-      );
-    }
-
-    const sessionResult = createOutboundSession({
-      channel: verificationChannel,
-      expectedExternalUserId: channel.externalUserId ?? undefined,
-      expectedChatId: channel.externalChatId ?? undefined,
-      identityBindingStatus: "bound",
-      destinationAddress: slackUserId,
-      verificationPurpose: "trusted_contact",
-    });
-
-    const slackBody = composeVerificationSlack(
-      GUARDIAN_VERIFY_TEMPLATE_KEYS.SLACK_CHALLENGE_REQUEST,
-      {
-        code: sessionResult.secret,
-        expiresInMinutes: Math.floor(SESSION_TTL_SECONDS / 60),
-      },
-    );
-
-    const now = Date.now();
-    const sendCount = 1;
-    updateSessionDelivery(sessionResult.sessionId, now, sendCount, null);
-    deliverVerificationSlack(slackUserId, slackBody, assistantId);
-
-    return Response.json({
-      ok: true,
-      verificationSessionId: sessionResult.sessionId,
-      expiresAt: sessionResult.expiresAt,
-      sendCount,
-    });
-  }
-
-  return httpError(
-    "BAD_REQUEST",
-    `Verification is not supported for channel type "${channel.type}"`,
-    400,
-  );
-}
-
-// ---------------------------------------------------------------------------
 // Route definitions
 // ---------------------------------------------------------------------------
 
@@ -698,16 +435,6 @@ export function contactRouteDefinitions(): RouteDefinition[] {
       policyKey: "contact-channels",
       handler: async ({ req, params }) =>
         handleUpdateContactChannel(req, params.contactChannelId),
-    },
-    {
-      endpoint: "contact-channels/:contactChannelId/verify",
-      method: "POST",
-      policyKey: "contact-channels",
-      handler: async ({ params, authContext }) =>
-        handleVerifyContactChannel(
-          params.contactChannelId,
-          authContext.assistantId,
-        ),
     },
   ];
 }
